requirements.txt
pillow
tk
networkx
matplotlib